requirements.txt
numpy
opencv-python
pre-commit
rclpy-intellisense
ros-std-msgs
sensor_msgs_py
